sysrdl2jinja/requirements.txt
antlr4-python3-runtime==4.9.3
colorama==0.4.4
importlib-metadata==4.8
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
systemrdl-compiler==1.21.0
zipp==3.6.0
